helpers/systemInfo: extract info fns
