+++ src/helpers/MiscFunctions.hpp
@@ -27,7 +27,6 @@ bool                                    isDirection(const char&);
 SWorkspaceIDName                        getWorkspaceIDNameFromString(const std::string&);
 std::optional<std::string>              cleanCmdForWorkspace(const std::string&, std::string);
 float                                   vecToRectDistanceSquared(const Vector2D& vec, const Vector2D& p1, const Vector2D& p2);
-void                                    logSystemInfo();
 std::string                             execAndGet(const char*);
 int64_t                                 getPPIDof(int64_t pid);
 std::expected<int64_t, std::string>     configStringToInt(const std::string&);
@@ -46,6 +45,7 @@ std::string                             deviceNameToInternalString(const std::st
 std::string                             getSystemLibraryVersion(const std::string& name);
 std::string                             getBuiltSystemLibraryNames();
 bool                                    truthy(const std::string& str);
+std::string                             systemInfo();
 
 template <typename... Args>
 [[deprecated("use std::format instead")]] std::string getFormat(std::format_string<Args...> fmt, Args&&... args) {
